Add catch error listener
